Rename webgl.rs to webgl_bindngs.rs
